Refactor morph and morph-rendering module
