+++ cocos/3d/assets/morph.ts
@@ -23,29 +23,13 @@
  THE SOFTWARE.
  */
 
-import { AttributeName, Device, DescriptorSet } from '../../core/gfx';
+import { AttributeName } from '../../core/gfx';
 import { Mesh } from './mesh';
-import { StdMorphRendering } from './morph-rendering';
-import { IMacroPatch } from '../../core/renderer';
 
-export interface Morph {
-    /**
-     * Morph data of each sub-mesh.
-     */
-    subMeshMorphs: (SubMeshMorph | null)[];
-
-    /**
-     * Common initial weights of each sub-mesh.
-     */
-    weights?: number[];
-
-    /**
-     * Name of each target of each sub-mesh morph.
-     * This field is only meaningful if every sub-mesh has the same number of targets.
-     */
-    targetNames?: string[];
-}
-
+/**
+ * @en Morph target contains all displacements data of each vertex attribute like position and normal.
+ * @zh 形变目标数据包含网格顶点属性在形变下的变化值，可能包含位移、法线等属性
+ */
 export interface MorphTarget {
     /**
      * Displacement of each target attribute.
@@ -53,6 +37,11 @@ export interface MorphTarget {
     displacements: Mesh.IBufferView[];
 }
 
+/**
+ * @en Sub mesh morph data describes all morph targets for one sub mesh,
+ * including attributes in each morph target, morph targets data and weights corresponding each targets.
+ * @zh 子网格形变数据描述一个子网格下所有的形变目标数据，包含顶点形变属性，形变目标数据和对应每个形变目标的权重。
+ */
 export interface SubMeshMorph {
     /**
      * Attributes to morph.
@@ -70,39 +59,28 @@ export interface SubMeshMorph {
     weights?: number[];
 }
 
-export function createMorphRendering (mesh: Mesh, gfxDevice: Device): MorphRendering | null {
-    return new StdMorphRendering(mesh, gfxDevice);
-}
-
 /**
- * Class which control rendering of a morph resource.
+ * @en Mesh morph data structure to describe the sub meshes data of all sub meshes,
+ * it also contains all sub mesh morphs, global weights configuration and target names.
+ * Normally the global weights configuration should be identical to the sub mesh morph weights,
+ * but if not, the global weights in morph is less prioritized.
+ * @zh 网格的形变数据结构，包含所有子网格形变数据，全局的权重配置和所有形变目标名称。
+ * 一般来说，全局权重配置和子网格形变数据中保持一致，但如果有差异，以子网格形变数据中的权重配置为准。
  */
-export interface MorphRendering {
-    createInstance (): MorphRenderingInstance;
-}
-
-/**
- * This rendering instance of a morph resource.
- */
-export interface MorphRenderingInstance {
+export interface Morph {
     /**
-     * Sets weights of targets of specified sub mesh.
-     * @param subMeshIndex
-     * @param weights
+     * Morph data of each sub-mesh.
      */
-    setWeights (subMeshIndex: number, weights: number[]): void;
+    subMeshMorphs: (SubMeshMorph | null)[];
 
     /**
-     * Adapts pipeline state to do the rendering.
-     * @param subMeshIndex
-     * @param pipelineState
+     * Common initial weights of each sub-mesh.
      */
-    adaptPipelineState (subMeshIndex: number, descriptorSet: DescriptorSet): void;
-
-    requiredPatches (subMeshIndex: number): IMacroPatch[] | null;
+    weights?: number[];
 
     /**
-     * Destroy the rendering instance.
+     * Name of each target of each sub-mesh morph.
+     * This field is only meaningful if every sub-mesh has the same number of targets.
      */
-    destroy (): void;
+    targetNames?: string[];
 }
