+++ cocos/3d/models/morph-model.ts
@@ -24,7 +24,7 @@
  */
 
 import { Model } from '../../core/renderer/scene/model';
-import { MorphRenderingInstance } from '../assets/morph';
+import { MorphRenderingInstance } from '../assets/morph-rendering';
 import { Material } from '../../core/assets/material';
 import { RenderingSubMesh } from '../../core/assets/rendering-sub-mesh';
 import { DescriptorSet } from '../../core/gfx';
